Throw away ParseFile::$tokens after parsing
This saves A LOT of ram. Memory useage dropped from ~400M to ~5M
for parsing of a large project (929 files).

Note that one of the debugging lines in relation to broken class refs
needed to be removed, as it was referencing the $tokens variable. This
error does not come up very often so hopefully this omission will not
cause major issues.